+++ Code2Gether-Discord-Bot.Tests/CreateProjectTests.cs
@@ -53,7 +53,7 @@ namespace Code2Gether_Discord_Bot.Tests
 
             _repo = new FakeProjectRepository()
             {
-                Projects = new Dictionary<long, Project>()
+                Projects = new Dictionary<int, Project>()
                 {
                     {0, new Project(0, "unittest", user)},
                 }
